+++ addons/config/options.lua
@@ -1,13 +1,13 @@
 --[[
-Copyright 2012-2023 João Cardoso
-All Rights Reserved
+	Copyright 2012-2023 João Cardoso
+	All Rights Reserved
 --]]
 
 local Sushi, Addon = LibStub('Sushi-3.2'), PetTracker
 local Options = PetTracker:NewModule('Options', Sushi.OptionsGroup('|Tinterface/addons/pettracker/art/compass:16:16|t PetTracker'))
 local L = LibStub('AceLocale-3.0'):GetLocale('PetTracker')
 
-local PATRONS = {{title='Jenkins',people={'Gnare','Adcantu','Justin Hall','Debora S Ogormanw','Johnny Rabbit','Francesco Rollo'}},{title='Ambassador',people={'Julia F','Lolari ','Dodgen','Kopernikus ','Ptsdthegamer','Burt Humburg','Adam Mann','Christie Hopkins','Bc Spear','Jury ','Tigran Andrew','Jeffrey Jones','Swallow@area52','Peter Hollaubek','Michael Kinasz','Sam Ramji','Kelly Wolf','Syed Hamdani','Thinkdesigner ','Charles Howarth','Harry J Hightower'}}} -- generated patron list
+local PATRONS = {{title='Jenkins',people={'Gnare','Adcantu','Justin Hall','Debora S Ogormanw','Johnny Rabbit','Francesco Rollo'}},{title='Ambassador',people={'Julia F','Lolari ','Dodgen','Kelly Wolf','Kopernikus ','Ptsdthegamer','Burt Humburg','Adam Mann','Christie Hopkins','Bc Spear','Jury ','Tigran Andrew','Swallow@area52','Peter Hollaubek','Michael Kinasz','Sam Ramji','Syed Hamdani','Ds9293','Charles Howarth'}}} -- generated patron list
 local PATREON_ICON = '  |TInterface/Addons/PetTracker/art/patreon:12:12|t'
 local HELP_ICON = '  |T516770:13:13:0:0:64:64:14:50:14:50|t'
 local FOOTER = 'Copyright 2012-2023 João Cardoso'
@@ -16,7 +16,7 @@ local FOOTER = 'Copyright 2012-2023 João Cardoso'
 --[[ Startup ]]--
 
 function Options:OnEnable()
-	self.FAQ = Sushi.OptionsGroup(self, HELP_LABEL .. HELP_ICON)
+	self.Help = Sushi.OptionsGroup(self, HELP_LABEL .. HELP_ICON)
 					:SetSubtitle(L.HelpDescription):SetFooter(FOOTER):SetChildren(self.OnHelp)
 	self.Credits = Sushi.OptionsGroup(self, 'Patrons' .. PATREON_ICON)
 					:SetSubtitle(L.PatronsDescription):SetFooter(FOOTER):SetOrientation('HORIZONTAL'):SetChildren(self.OnCredits)
@@ -66,6 +66,7 @@ function Options:OnCredits()
 		end
 	end
 
+	self:AddBreak()
 	self:Add('RedButton', 'Join Us'):SetWidth(200):SetCall('OnClick', function()
 		Sushi.Popup:External('patreon.com/jaliborc')
 		SettingsPanel:Close(true)