+++ addons/main/main.lua
@@ -4,6 +4,7 @@ All Rights Reserved
 --]]
 
 local ADDON, Addon = ...
+local C = LibStub('C_Everywhere')
 local Addon = LibStub('WildAddon-1.0'):NewAddon(ADDON, Addon, 'MutexDelay-1.0')
 Addon.MaxQuality = 6
 Addon.MaxLevel = 25
@@ -12,23 +13,32 @@ Addon.MaxLevel = 25
 --[[ Events ]]--
 
 function Addon:OnEnable()
-	PetTracker_State = PetTracker_State or {}
-	PetTracker_Sets = setmetatable(PetTracker_Sets or {}, {__index = {
+	LibStub('LibPetJournal-2.0').RegisterCallback(self, 'PostPetListUpdated', 'OnPetsChanged')
+	self:RegisterEvent('PET_BATTLE_OPENING_START', 'OnBattle')
+	self.state = PetTracker_State or {}
+	self.sets = setmetatable(PetTracker_Sets or {}, {__index = {
 		zoneTracker = true, capturedPets = true, specieIcons = true, rivalPortraits = true,
 		switcher = true, alertUpgrades = true, forfeit = true,
 	}})
 
-	self.sets, self.state = PetTracker_Sets, PetTracker_State
 	if self.sets.tutorial == 12 then
-		CreateFrame('Frame', nil, SettingsPanel or InterfaceOptionsFrame):SetScript('OnShow', function()
-			LoadAddOn(ADDON .. '_Config')
+		(SettingsPanel or InterfaceOptionsFrame):HookScript('OnShow', function()
+			C.AddOns.LoadAddOn(ADDON .. '_Config')
 		end)
 	else
-		LoadAddOn(ADDON .. '_Config')
+		C.AddOns.LoadAddOn(ADDON .. '_Config')
 	end
 
-	LibStub('LibPetJournal-2.0').RegisterCallback(self, 'PostPetListUpdated', 'OnPetsChanged')
-	self:RegisterEvent('PET_BATTLE_OPENING_START', 'OnBattle')
+	PetTracker_Sets, PetTracker_State = self.sets, self.state
+	AddonCompartmentFrame:RegisterAddon {
+		text = ADDON, keepShownOnClick = true, notCheckable = true,
+		icon = 'interface/addons/pettracker/art/compass',
+		func = function()
+			if C.AddOns.LoadAddOn(ADDON .. '_Config') then
+				Addon.Options:Open()
+			end
+		end
+	}
 end
 
 function Addon:OnPetsChanged()
@@ -36,7 +46,7 @@ function Addon:OnPetsChanged()
 end
 
 function Addon:OnBattle()
-	if LoadAddOn(ADDON .. '_Battle') then
+	if C.AddOns.LoadAddOn(ADDON .. '_Battle') then
 		self:SendSignal('BATTLE_STARTED')
 	end
 end